// GenericCmsClient/src/app/api/model/autocompleteDataProviderModelEntityQuery.ts
/**
 * GenericCms
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AutocompleteDataProviderModelEntityQuery { 
    readonly type?: string | null;
    name?: string | null;
    expression?: any | null;
    property?: string | null;
}